Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "passlib>=1.7.4",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.47",
]
